Add files via upload
This is a starting point assuming a base start. Everything here can be completed before the base, though some things require story progression.  Some side quests may be too involved for a 45 minute match, but we can discuss.  All included machines are present and killable outside of a quest or contracts.  I didn't include burrowers or scroungers because that's what we kill in the prologue, and the number killed can vary.